@@ -20,7 +20,8 @@
     }
   },
   "dependencies": {
-    "@vue/component-compiler-utils": "^3.1.0",
+    "@mdi/font": "^7.4.47",
+    "@vue/compiler-sfc": "^3.5.18",
     "axios": "^0.19.0",
     "axios-debug-log": "^0.8.0",
     "cheerio": "^1.0.0-rc.3",
@@ -30,20 +31,21 @@
     "firebase": "^7.10.0",
     "interval-call": "^0.4.0",
     "jquery": "^3.4.1",
+    "lodash": "^4.17.21",
     "md5": "^2.2.1",
     "qs": "^6.9.4",
     "readability-js": "^1.0.7",
     "sass": "^1.24.0",
+    "sortablejs-vue3": "^1.2.11",
     "sweetalert2": "^9.10.12",
-    "vue": "^2.6.11",
+    "vue": "^3.5.18",
     "vue-drag-drop": "^1.1.4",
-    "vue-lodash": "^2.0.2",
-    "vue-router": "^3.0.1",
+    "vue-loader": "^17.4.2",
+    "vue-lodash": "^2.1.2",
+    "vue-router": "^4.5.1",
     "vue-style-loader": "^4.1.2",
-    "vuedraggable": "^2.24.0",
-    "vuetify": "^2.1.15",
-    "vuetify-loader": "^1.4.3",
-    "vuex": "^3.0.1",
+    "vuetify": "^3.9.4",
+    "vuex": "^4.1.0",
     "webextension-polyfill": "^0.3.1"
   },
   "devDependencies": {
@@ -51,6 +53,7 @@
     "@babel/plugin-proposal-optional-chaining": "^7.0.0",
     "@babel/preset-env": "^7.1.0",
     "@babel/runtime-corejs3": "^7.4.0",
+    "@vitejs/plugin-vue": "^6.0.1",
     "archiver": "^3.0.0",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.2",
@@ -73,12 +76,13 @@
     "file-loader": "^1.1.11",
     "husky": "^2.4.0",
     "mini-css-extract-plugin": "^0.4.4",
-    "node-sass": "^4.9.3",
     "prettier": "^1.17.1",
     "pretty-quick": "^1.8.0",
+    "sass": "^1.24.0",
     "sass-loader": "^7.3.1",
-    "vue-loader": "^15.4.2",
-    "vue-template-compiler": "^2.6.11",
+    "vite": "^7.1.1",
+    "vite-plugin-static-copy": "^3.1.1",
+    "vite-plugin-web-extension": "^4.4.5",
     "web-ext-types": "^2.1.0",
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.2",
