@@ -1,9 +1,7 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App";
 
 global.browser = require("webextension-polyfill");
-/* eslint-disable no-new */
-new Vue({
-  el: "#app",
-  render: h => h(App)
-});
+
+const app = createApp(App);
+app.mount("#app");
