@@ -1,10 +1,10 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import DetectPopupApp from "./DetectPopupApp";
 import vuetify from "../lib/vuetify/vuetify";
 
-/* eslint-disable no-new */
-new Vue({
-  el: "#app",
-  vuetify,
-  render: h => h(DetectPopupApp)
-});
+// Import Vuetify compatibility CSS
+import "../css/vuetify-compat.css";
+
+const app = createApp(DetectPopupApp);
+app.use(vuetify);
+app.mount("#app");
