@@ -1,4 +1,33 @@
 // event-bus.js
-import Vue from "vue";
-const EventBus = new Vue();
-export default EventBus;
+import { reactive } from "vue";
+
+class EventBus {
+  constructor() {
+    this.events = reactive({});
+  }
+
+  $emit(event, ...args) {
+    if (this.events[event]) {
+      this.events[event].forEach(callback => callback(...args));
+    }
+  }
+
+  $on(event, callback) {
+    if (!this.events[event]) {
+      this.events[event] = [];
+    }
+    this.events[event].push(callback);
+  }
+
+  $off(event, callback) {
+    if (this.events[event]) {
+      const index = this.events[event].indexOf(callback);
+      if (index > -1) {
+        this.events[event].splice(index, 1);
+      }
+    }
+  }
+}
+
+const eventBus = new EventBus();
+export default eventBus;
