@@ -1,47 +1,40 @@
 <template>
-  <v-app-bar app clipped-left color="">
-    <!--<v-app-bar-nav-icon @click="drawer = !drawer"/>-->
+  <header class="app-bar app clipped-left">
+    <!--<button class="app-bar-nav-icon" @click="drawer = !drawer"/>-->
 
     <span class="title ml-3 mr-5" @click="processTest">
       WEBGALPI (
       <span style="color: #ff8b20;"
-        ><v-icon color="#ff8b20" style="width: 12px;">mdi-beta</v-icon>eta
+        ><span class="icon-beta" style="color: #ff8b20; font-size: 12px;"
+          >β</span
+        >eta
       </span>
       <span style="font-size: 12px">Ver.{{ version }}</span>
       )
     </span>
-    <v-text-field
-      solo-inverted
-      flat
-      hide-details
-      label="Search"
-      prepend-inner-icon="mdi-feature-search-outline"
+    <input
+      class="text-field solo-inverted flat hide-details"
+      placeholder="Search"
       v-if="false"
     />
-    <v-spacer />
+    <div class="spacer"></div>
 
-    <!-- <v-menu
-      transition="slide-y-transition"
-      offset-y
-      :close-on-content-click="false"
-    >
-      <template v-slot:activator="{ on }">
-        <v-btn text small v-on="on">
-          {{ member.EMAIL }}
-        </v-btn>
-      </template>
-      <v-list class="pt-0 pb-0">
-        <v-list-item class="pt-0 pb-0" @click="signOut">
+    <!-- <div class="menu">
+      <button class="btn text small">
+        {{ member.EMAIL }}
+      </button>
+      <ul class="menu-list pt-0 pb-0">
+        <li class="menu-item pt-0 pb-0" @click="signOut">
           {{ LANG.BUTTON_MESSAGE("B0002") }}
-        </v-list-item>
-      </v-list>
-    </v-menu>-->
+        </li>
+      </ul>
+    </div>-->
 
     &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
     <OptionPage ref="optionPage"></OptionPage>
 
     <SignArea ref="signout"></SignArea>
-  </v-app-bar>
+  </header>
 </template>
 <script>
 import SignArea from "../dialog/setting/SignArea";
@@ -85,3 +78,137 @@ export default {
   }
 };
 </script>
+<style>
+.app-bar {
+  display: flex;
+  align-items: center;
+  padding: 0 16px;
+  height: 64px;
+  background: #1976d2;
+  color: white;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.app-bar.app {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 1000;
+}
+
+.app-bar.clipped-left {
+  left: 0;
+}
+
+.title {
+  font-size: 1.25rem;
+  font-weight: 500;
+  cursor: pointer;
+}
+
+.text-field {
+  padding: 8px 12px;
+  border: none;
+  border-radius: 4px;
+  background: rgba(255, 255, 255, 0.1);
+  color: white;
+  min-width: 200px;
+}
+
+.text-field::placeholder {
+  color: rgba(255, 255, 255, 0.7);
+}
+
+.text-field.solo-inverted {
+  background: rgba(255, 255, 255, 0.1);
+}
+
+.text-field.flat {
+  box-shadow: none;
+}
+
+.text-field.hide-details {
+  margin-bottom: 0;
+}
+
+.spacer {
+  flex: 1;
+}
+
+.menu {
+  position: relative;
+  display: inline-block;
+}
+
+.menu-list {
+  display: none;
+  position: absolute;
+  top: 100%;
+  right: 0;
+  background: white;
+  color: black;
+  border-radius: 4px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  min-width: 150px;
+}
+
+.menu:hover .menu-list {
+  display: block;
+}
+
+.menu-item {
+  padding: 8px 16px;
+  cursor: pointer;
+}
+
+.menu-item:hover {
+  background: #f5f5f5;
+}
+
+.btn {
+  padding: 6px 12px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 14px;
+}
+
+.btn.text {
+  background: none;
+  color: inherit;
+}
+
+.btn.small {
+  padding: 4px 8px;
+  font-size: 12px;
+}
+
+.ml-3 {
+  margin-left: 24px;
+}
+
+.mr-5 {
+  margin-right: 40px;
+}
+
+.pt-0 {
+  padding-top: 0;
+}
+
+.pb-0 {
+  padding-bottom: 0;
+}
+
+.icon-beta {
+  display: inline-block;
+}
+
+/* Dark theme styles */
+body.theme-dark .app-bar {
+  background: #212121;
+}
+</style>
