@@ -1,83 +1,97 @@
 import { DDL } from "./DDL_Query";
 import Utils from "../dashboard/utils/Utils";
 
-let db = openDatabase("HL", "1.0", "DATABASE", 200000);
+// WebSQL is deprecated and not supported in Chrome extensions
+// Using chrome.storage.local instead
 
 export default {
   initData: accountInfo => {
     return new Promise(res => {
-      let INIT_OPTIONS =
-        `INSERT INTO TBL_OPTIONS  VALUES(\'` +
-        accountInfo.email +
-        `\','highlight-color-1,highlight-color-2,highlight-color-3,highlight-color-4,highlight-color-5,highlight-color-6' --COLOR
-                        ,'EN'
-                        ,'Y'--MESSAGE_ALTER
-                        ,'Y' --highlight
-                        ,'N' --rootsite
-                        ,'N' --excute_iframe
-                        ,'N' --viewWidget
-                        ,0
-                        ,null --EXCLUDE_SITE
-                        ,'N' --SYNC_BOOKMARK
-                        ,'dark' --THEME (dark / light);
-                    )`;
-
-      let INIT_CATEGORY =
-        `INSERT INTO TBL_CATEGORY(
-                                    IDX,
-                                     EMAIL,
-                                     NAME,
-                                     PARENT,
-                                     DEPTH,
-                                     SORT,
-                                     TYPE,
-                                     FLAG,
-                                     DATE_CREATE
-                            ) VALUES  (1,\'` +
-        accountInfo.email +
-        `\', 'DEFAULT CATEGORY', 0, 0, 0, 'SYSTEM','root', null),
-                                          (5,\'` +
-        accountInfo.email +
-        `\', 'ALL CATEGORY', 1, 1, 1, 'SYSTEM','all', null),
-                                          (6,\'` +
-        accountInfo.email +
-        `\', 'NO CATEGORY', 1, 1, 2, 'SYSTEM','nocategory', null)
-                                          `;
-
-      db.transaction(tx => {
-        tx.executeSql(INIT_OPTIONS, []);
-        tx.executeSql(INIT_CATEGORY, []);
-      });
-      res(true);
-    });
-  },
-  selectData: query => {
-    return new Promise(res => {
-      db.transaction(tx => {
-        tx.executeSql(query, null, (tx, rs) => {
-          let retArray = new Array();
-          for (let i = 0; i < rs.rows.length; i++) {
-            retArray.push(rs.rows.item(i));
+      // Initialize data using chrome.storage.local
+      if (accountInfo && accountInfo.email) {
+        chrome.storage.local.set(
+          {
+            currentUser: accountInfo.email,
+            userProfile: accountInfo
+          },
+          () => {
+            if (chrome.runtime.lastError) {
+              console.error("Storage error:", chrome.runtime.lastError);
+              res(false);
+            } else {
+              console.log("User data initialized:", accountInfo.email);
+              res(true);
+            }
           }
-          res(retArray);
-        });
+        );
+      } else {
+        res(true);
+      }
+    });
+  },
+  // selectData function removed - use direct chrome.storage.local operations instead
+  createTable: () => {
+    // Chrome extensions use chrome.storage.local - no table creation needed
+    console.log(
+      "Storage ready - no table creation needed for chrome.storage.local"
+    );
+    return Promise.resolve(true);
+  },
+  dropTable: tableName => {
+    return new Promise(res => {
+      // Remove specific data from chrome.storage.local
+      chrome.storage.local.remove([tableName], () => {
+        if (chrome.runtime.lastError) {
+          console.error("Storage remove error:", chrome.runtime.lastError);
+          res(false);
+        } else {
+          console.log("Storage data cleared:", tableName);
+          res(true);
+        }
       });
     });
   },
-  createTable: () => {
-    DDL.CREATE(db);
+  addTable: (tableName, data) => {
+    return new Promise(res => {
+      // Add data to chrome.storage.local
+      let storageData = {};
+      storageData[tableName] = data;
+
+      chrome.storage.local.set(storageData, () => {
+        if (chrome.runtime.lastError) {
+          console.error("Storage add error:", chrome.runtime.lastError);
+          res(false);
+        } else {
+          console.log("Storage data added:", tableName);
+          res(true);
+        }
+      });
+    });
   },
-  dropTable: () => {
-    DDL.DROP(db);
+  removeTable: tableName => {
+    return new Promise(res => {
+      chrome.storage.local.remove([tableName], () => {
+        if (chrome.runtime.lastError) {
+          console.error("Storage remove error:", chrome.runtime.lastError);
+          res(false);
+        } else {
+          console.log("Storage table removed:", tableName);
+          res(true);
+        }
+      });
+    });
   },
-  addTable: () => {
-    DDL.ADD_TABLE(db);
-  },
-  removeTable: () => {
-    DDL.DROP_TABLE(db);
-  },
-  truncateTable: async () => {
-    let result = await Utils.getLocalStorage("loginInfo");
-    DDL.TRUNCATE(db, result.loginInfo);
+  truncateTable: async tableName => {
+    return new Promise(res => {
+      chrome.storage.local.remove([tableName], () => {
+        if (chrome.runtime.lastError) {
+          console.error("Storage truncate error:", chrome.runtime.lastError);
+          res(false);
+        } else {
+          console.log("Storage table truncated:", tableName);
+          res(true);
+        }
+      });
+    });
   }
 };
